fix(training): continued runs report cumulative epochs; scope epoch filenames by version
Continued ("train further") training restarted its epoch numbering at 1, so a
continuation reported epochs 1-10 again and its downloads collided with the
source run's. Two separate defects behind one report:

Numbering: `trainingResults.epochOffset` is stamped once, at submit, from
`continueFromEpoch.epochNumber`, and only read at ingest. Deriving it at ingest
would renumber a FINISHED run on any re-sync, and generation workflows bind an
epoch by value — a lookup miss resolves to the newest epoch rather than erroring,
so those bindings would silently move to other weights. Runs predating the field
have no offset and keep the numbers they already have.

Filenames: the collision was wider than continuations — every training run gets
its own model version and numbers epochs from its own start, so two independent
runs on one model always produced `<model>_epoch_1..N` twice. All four naming
sites now go through one helper and are scoped by version. The version id is
what makes that a guarantee: 496 trained models have two versions sharing a
name. Names are also capped, which fixes 42 versions whose filenames already
exceeded the 255-byte path-component limit.

Progress readouts now report `highest epoch - offset` rather than the raw
highest number (inflated for continuations) or the epoch count (wrong for the
~5% of runs that save checkpoints at a stride — a finished 50-epoch Kohya run
commonly stores 18). Verified a no-op across 200k existing runs.

Also adds Axiom logging to the training what-if, which previously failed with
no server-side trace behind a bare "Error computing cost".

+++ src/server/schema/model-version.schema.ts
@@ -263,7 +263,9 @@ export const trainingDetailsObj = z.object({
   continueFromEpoch: z
     .object({
       air: z.string(),
-      epochNumber: z.number(),
+      // Not `.nonnegative()`: an epoch the orchestrator never numbered is stored as -1, and
+      // "Train Further" sends that value straight back. The service clamps it instead.
+      epochNumber: z.number().int(),
       sourceModelVersionId: z.number(),
       sourceVersionName: z.string().optional(),
     })
